requirements.txt
streamlit
pypdf
python-docx
groq
python-dotenv
